+++ packages/moonstone/CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Changed
+
+- `moonstone/ExpandableList` to accept an array of objects as children which are spread onto the generated components
+
 ### Fixed
 
 - `moonstone/Panels.Panel` behavior for setting focus after render